+++ app/gui/src/dashboard/providers/LocalStorageProvider.tsx
@@ -31,7 +31,8 @@ export type LocalStorageProviderProps = Readonly<React.PropsWithChildren>
 /** A React Provider that lets components get the shortcut registry. */
 export default function LocalStorageProvider(props: LocalStorageProviderProps) {
   const { children } = props
-  const [localStorage] = React.useState(() => new LocalStorage())
+
+  const localStorage = React.useMemo(() => LocalStorage.getInstance(), [])
 
   return (
     <LocalStorageContext.Provider value={{ localStorage }}>{children}</LocalStorageContext.Provider>
@@ -86,6 +87,14 @@ export function useLocalStorageState<K extends LocalStorageKey>(
         return nextValue
       })
     },
+  )
+
+  React.useEffect(
+    () =>
+      localStorage.subscribe(key, (newValue) => {
+        privateSetValue(newValue ?? defaultValue)
+      }),
+    [defaultValue, key, localStorage],
   )
 
   return [value, setValue]