+++ app/gui/src/project-view/bindings.ts
@@ -12,6 +12,8 @@ export const codeEditorBindings = defineKeybinds('code-editor', {
 export const documentationEditorBindings = defineKeybinds('documentation-editor', {
   toggle: ['Mod+D'],
   openLink: ['Mod+PointerMain'],
+  paste: ['Mod+V'],
+  pasteRaw: ['Mod+Shift+V'],
 })
 
 export const interactionBindings = defineKeybinds('current-interaction', {